17 bug issue the market tool stopped working following the mw2 and mw3 updates

+++ LINDRA - Market/MarketOffsets/Interfaces/ISafeArea.cs
@@ -9,8 +9,8 @@ namespace MarketOffsets.Interfaces
     internal interface ISafeArea
     {
 
-        int profileMenuOption_safeAreaHorz { get; }
-        int profileMenuOption_safeAreaVert { get; }
+        long profileMenuOption_safeAreaHorz { get; }
+        long profileMenuOption_safeAreaVert { get; }
 
     }
 }
